+++ src/lsp/mod.rs
@@ -9,6 +9,7 @@ pub(crate) mod hover;
 pub(crate) mod inlay_hints;
 pub(crate) mod navigation;
 pub(crate) mod renaming;
+pub(crate) mod scope;
 pub(crate) mod selection_ranges;
 pub(crate) mod semantic_tokens;
 pub(crate) mod signature_help;